feat(Section/Users/Modal): checking the teacher role to display the availability modal
