+++ components/Section/Usuarios/Form/FormUserProfile.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
-import { useForm } from 'vee-validate';
-import * as yup from 'yup';
+import { useFormUser, type FormUserOutput } from './FormUser';
 
 const props = defineProps({
 	//props do modal criar e editar
@@ -15,70 +14,9 @@ const editIdRef = toRef(props, 'editId');
 
 const $emit = defineEmits(['close']);
 
-type FormOutput = {
-	imagem: Blob | null | undefined;
+const { resetForm, handleSubmit } = useFormUser();
 
-	nome: string;
-
-	codigo: string;
-};
-
-type FormValues = {
-	imagem: Blob | null | undefined;
-	nome: string;
-	email: string;
-
-	vinculos: Array<{
-		campus: { id: string };
-		cargos: Array<'dape' | 'professor'>;
-	}>;
-};
-
-const initialFormValues = reactive({
-	imagem: null,
-	nome: '',
-	codigo: '',
-	vinculos: [
-		{
-			campus: { id: null } as any,
-			cargos: [] as any,
-		},
-	] as any,
-});
-
-const schema = yup.object().shape({
-	imagem: yup.mixed().nullable().optional(),
-
-	nome: yup.string().required('Nome é obrigatório!'),
-
-	email: yup.string().required('Email é obrigatório!'),
-
-	vinculos: yup.array().of(
-		yup.object({
-			campus: yup.object({
-				id: yup.string().required('Informe o campus deste vínculo!'),
-			}),
-			cargos: yup
-				.array()
-				.of(yup.string())
-				.min(
-					1,
-					'O usuário deve possuir ao menos 1 cargo neste vínculo!'
-				),
-		})
-	),
-});
-
-const {
-	resetForm,
-	handleSubmit,
-	values: formValues,
-} = useForm<FormValues, FormOutput>({
-	validationSchema: schema,
-	initialValues: initialFormValues,
-});
-
-const onSubmit = handleSubmit(async (values: FormOutput) => {
+const onSubmit = handleSubmit(async (values: FormUserOutput) => {
 	const editId = editIdRef.value;
 
 	const { imagem, ...data } = values;
@@ -93,8 +31,8 @@ const onSubmit = handleSubmit(async (values: FormOutput) => {
 		<div class="modal">
 			<div class="form-header">
 				<h1 class="main-title">
-					<span v-if="editId">Editar Turma</span>
-					<span v-else>Cadastrar Nova Turma</span>
+					<span v-if="editId">Editar Usuário</span>
+					<span v-else>Cadastrar Usuário</span>
 				</h1>
 			</div>
