--- components/Section/Usuarios/ModalNewUser/Funcao.vue
@@ -1,13 +1,0 @@
-<template>
-	<SelectBase
-		label="Função"
-		placeholder="Selecione"
-		:items="['Professor', 'DAPE']"
-		multiple
-		:clearable="false"
-	/>
-</template>
-
-<script setup>
-import SelectBase from '~/components/UI/SelectBase/SelectBase.vue';
-</script>
