+++ pages/dashboard/usuarios/index.vue
@@ -3,14 +3,3 @@
 		<SectionUsuarios />
 	</div>
 </template>
-
-<script lang="ts" setup></script>
-
-<style>
-.welcomeMessage {
-	font-family: 'Poppins', sans-serif;
-	font-size: 1.5rem;
-	font-weight: 600;
-	color: #4f4f4f;
-}
-</style>
